% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitkatfunction.R
\name{acube}
\alias{acube}
\title{Area of cube}
\usage{
acube(r)
}
\arguments{
\item{r}{is the square side length}
}
\value{
The area of cube
}
\description{
Compute the area of cube shape
}
\examples{
area1 <- acube(4);
area2 <- acube(3);
}
