% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitkatfunction.R
\name{acylinder}
\alias{acylinder}
\title{Area of cylinder}
\usage{
acylinder(r, h)
}
\arguments{
\item{r}{is the radius of the cylinder}

\item{h}{is the height of the cylinder}
}
\value{
The area of cylinder
}
\description{
Compute the area of cylinder shape
}
\examples{
area1 <- acylinder(4);
area2 <- acylinder(3);
}
