% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitkatfunction.R
\name{psemicircle}
\alias{psemicircle}
\title{Perimeter of semicircle}
\usage{
psemicircle(r)
}
\arguments{
\item{r}{is the radius of the semicircle}
}
\value{
The perimeter of semicircle
}
\description{
Compute the perimeter of semicircle shape
}
\examples{
perimeter1 <- psemicircle(5);
perimeter2 <- psemicircle(8);
}
