% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitkatfunction.R
\name{vcylinder}
\alias{vcylinder}
\title{Volume of cylinder (tube)}
\usage{
vcylinder(r, T)
}
\arguments{
\item{r}{is the radius of the cirlce}

\item{T}{is the height of cylinder}
}
\value{
Volume of cylinder (tube)
}
\description{
Compute the area of circle shape
}
\examples{
volume1 <- vcylinder(4,2);
volume2 <- vcylinder (3,1);
}
