% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitkatfunction.R
\name{vrectangularprism}
\alias{vrectangularprism}
\title{Volume of rectangular prism}
\usage{
vrectangularprism(l, w, h)
}
\arguments{
\item{l}{is the base length}

\item{w}{is the base width}

\item{h}{is the height of the prism}
}
\value{
The volume rectangular of prism
}
\description{
Compute the volume of rectangular prism shape
}
\examples{
volume1 <- vrectangularprism(2,1,4);
volume2 <- vrectangularprism(4,5,1);
}
