% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kitkatfunction.R
\name{vtriangularpyramid}
\alias{vtriangularpyramid}
\title{Volume of triangular pyramid}
\usage{
vtriangularpyramid(a, t, T)
}
\arguments{
\item{a}{is the area of the base}

\item{t}{is the height of triangle}

\item{T}{is the height of triangular prism}
}
\value{
The volume triangular of pyramid
}
\description{
Compute the volume of triangular pyramid shape
}
\examples{
volume1 <- vpyramid(5,3,2);
volume2 <- vpyramid(4,3,1);
}
